% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_offset_table.R
\name{get_offset_table}
\alias{get_offset_table}
\title{Calculates offset table for number of maximum overlapping positions}
\usage{
get_offset_table(max_positions, offset_x, offset_y)
}
\arguments{
\item{max_positions}{number of maximal exact overlaps}

\item{offset_x}{offset for positon distance}

\item{offset_y}{offset for in-between layer distance}
}
\value{
data frame with position, offsets_x and offsets_y
}
\description{
Calculates offset table for number of maximum overlapping positions
}
