% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MusicianInterests-data.R
\docType{data}
\name{MusicianInterests}
\alias{MusicianInterests}
\title{Survey about genre interests of some hobby musicians}
\format{An object of class \code{"data.frame"};}
\usage{
data(MusicianInterests)
}
\description{
Tibble of what genre they are interested in, what instrument they play
and what level the play their instrument at (1 = beginner, 2 = intermediate,
3 = experienced, 4 = very experienced, 5 = pro). Also there is an ID for
the musician.
}
\examples{
library(ggBubbles)
data(MusicianInterests)
head(MusicianInterests)
}
\keyword{datasets}
