% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MusicianInterestsSmall-data.R
\docType{data}
\name{MusicianInterestsSmall}
\alias{MusicianInterestsSmall}
\title{Small test data of musician, interest and experience study}
\format{An object of class \code{"data.frame"};}
\usage{
data(MusicianInterestsSmall)
}
\description{
Data.frame of what genre they are interested in, what instrument they play
and what level the play their instrument at.
}
\examples{
library(ggBubbles)
data(MusicianInterestsSmall)
head(MusicianInterestsSmall)
}
\keyword{datasets}
